refactor: separate ActiveSession
Move com.appland.appmap.record.ActiveSession into its own file.

Also, fix a lint warning in SqlQuery.
